[BE] chore : mino에 사용되는 Block 클래스를 별도 파일 분리 #22
