+++ server/src/service/minos/mino.ts
@@ -1,24 +1,7 @@
-import Position from "@/types/position";
 import STATUS from "@/service/status";
+import Block from "@/service/minos/block";
 
 const TETRA = 4;
-
-class Block {
-    color?: string;
-    xPos: number;
-    yPos: number;
-
-    constructor(color?: string, xPos?: number, yPos?: number) {
-        this.color = color;
-        this.xPos = xPos || 0;
-        this.yPos = yPos || 0;
-    }
-
-    setPosition({ xPos, yPos }: Position) {
-        this.xPos = xPos;
-        this.yPos = yPos;
-    }
-}
 
 class Mino {
     protected blockArea: number[][];
